Move io wrappers into impl/io.h for easy sharing
